Sync with upstream

+++ src/webui/httpconnection.cpp
@@ -166,7 +166,7 @@ void HttpConnection::translateDocument(QString& data) {
 
       QString translation = word;
       if (isTranslationNeeded) {
-        int context_index = 0;
+        size_t context_index = 0;
         while(context_index < context_count && translation == word) {
 #if (QT_VERSION < QT_VERSION_CHECK(5, 0, 0))
           translation = qApp->translate(contexts[context_index].c_str(), word.constData(), 0, QCoreApplication::UnicodeUTF8, 1);
